+++ .pnp.cjs
@@ -2447,8 +2447,8 @@ const RAW_RUNTIME_STATE =
       "npm:3.6.0"\
     ],\
     [\
-      "readonly-date",\
-      "npm:1.0.0"\
+      "readonly-date-esm",\
+      "npm:2.0.0"\
     ],\
     [\
       "rechoir",\
@@ -3325,7 +3325,7 @@ const RAW_RUNTIME_STATE =
           ["karma-jasmine-html-reporter", "virtual:4f1584ad4aba8733a24be7c8aebbffafef25607f2d00f4b314cf96717145c692763628a31c2b85d4686fbb091ff21ebffa3cc337399c042c19a32b9bdb786464#npm:1.6.0"],\
           ["nyc", "npm:17.1.0"],\
           ["prettier", "npm:3.6.2"],\
-          ["readonly-date", "npm:1.0.0"],\
+          ["readonly-date-esm", "npm:2.0.0"],\
           ["ses", "npm:1.13.1"],\
           ["source-map-support", "npm:0.5.21"],\
           ["typedoc", "virtual:4f1584ad4aba8733a24be7c8aebbffafef25607f2d00f4b314cf96717145c692763628a31c2b85d4686fbb091ff21ebffa3cc337399c042c19a32b9bdb786464#npm:0.28.5"],\
@@ -3414,7 +3414,7 @@ const RAW_RUNTIME_STATE =
           ["karma-jasmine-html-reporter", "virtual:4f1584ad4aba8733a24be7c8aebbffafef25607f2d00f4b314cf96717145c692763628a31c2b85d4686fbb091ff21ebffa3cc337399c042c19a32b9bdb786464#npm:1.6.0"],\
           ["nyc", "npm:17.1.0"],\
           ["prettier", "npm:3.6.2"],\
-          ["readonly-date", "npm:1.0.0"],\
+          ["readonly-date-esm", "npm:2.0.0"],\
           ["ses", "npm:1.13.1"],\
           ["source-map-support", "npm:0.5.21"],\
           ["typedoc", "virtual:4f1584ad4aba8733a24be7c8aebbffafef25607f2d00f4b314cf96717145c692763628a31c2b85d4686fbb091ff21ebffa3cc337399c042c19a32b9bdb786464#npm:0.28.5"],\
@@ -3686,7 +3686,7 @@ const RAW_RUNTIME_STATE =
           ["karma-jasmine-html-reporter", "virtual:4f1584ad4aba8733a24be7c8aebbffafef25607f2d00f4b314cf96717145c692763628a31c2b85d4686fbb091ff21ebffa3cc337399c042c19a32b9bdb786464#npm:1.6.0"],\
           ["nyc", "npm:17.1.0"],\
           ["prettier", "npm:3.6.2"],\
-          ["readonly-date", "npm:1.0.0"],\
+          ["readonly-date-esm", "npm:2.0.0"],\
           ["ses", "npm:1.13.1"],\
           ["source-map-support", "npm:0.5.21"],\
           ["typedoc", "virtual:4f1584ad4aba8733a24be7c8aebbffafef25607f2d00f4b314cf96717145c692763628a31c2b85d4686fbb091ff21ebffa3cc337399c042c19a32b9bdb786464#npm:0.28.5"],\
@@ -3757,7 +3757,7 @@ const RAW_RUNTIME_STATE =
           ["karma-jasmine-html-reporter", "virtual:4f1584ad4aba8733a24be7c8aebbffafef25607f2d00f4b314cf96717145c692763628a31c2b85d4686fbb091ff21ebffa3cc337399c042c19a32b9bdb786464#npm:1.6.0"],\
           ["nyc", "npm:17.1.0"],\
           ["prettier", "npm:3.6.2"],\
-          ["readonly-date", "npm:1.0.0"],\
+          ["readonly-date-esm", "npm:2.0.0"],\
           ["ses", "npm:1.13.1"],\
           ["source-map-support", "npm:0.5.21"],\
           ["typedoc", "virtual:4f1584ad4aba8733a24be7c8aebbffafef25607f2d00f4b314cf96717145c692763628a31c2b85d4686fbb091ff21ebffa3cc337399c042c19a32b9bdb786464#npm:0.28.5"],\
@@ -11426,11 +11426,11 @@ const RAW_RUNTIME_STATE =
         "linkType": "HARD"\
       }]\
     ]],\
-    ["readonly-date", [\
-      ["npm:1.0.0", {\
-        "packageLocation": "./.yarn/cache/readonly-date-npm-1.0.0-e93ba2496b-7ab32bf19f.zip/node_modules/readonly-date/",\
+    ["readonly-date-esm", [\
+      ["npm:2.0.0", {\
+        "packageLocation": "./.yarn/cache/readonly-date-esm-npm-2.0.0-bd68df4d29-31c21d0b54.zip/node_modules/readonly-date-esm/",\
         "packageDependencies": [\
-          ["readonly-date", "npm:1.0.0"]\
+          ["readonly-date-esm", "npm:2.0.0"]\
         ],\
         "linkType": "HARD"\
       }]\