+++ packages/tendermint-rpc/src/tendermint37/hasher.spec.ts
@@ -1,5 +1,5 @@
 import { fromBase64, fromHex } from "@cosmjs/encoding";
-import { ReadonlyDate } from "readonly-date";
+import { ReadonlyDate } from "readonly-date-esm";
 
 import { ReadonlyDateWithNanoseconds } from "../dates";
 import { hashBlock, hashTx } from "./hasher";